--- ts-sdk/actions/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@orca-so/whirlpools-actions",
-  "version": "0.0.2",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    "import": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    },
-    "require": {
-      "types": "./dist/index.d.cts",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts --sourcemap",
-    "clean": "rimraf dist"
-  },
-  "dependencies": {
-    "@orca-so/tx-sender": "*",
-    "@orca-so/whirlpools": "*"
-  },
-  "peerDependencies": {
-    "@solana/kit": "^2.1.0"
-  },
-  "devDependencies": {
-    "@solana/kit": "^2.1.0",
-    "tsup": "^8.3.6"
-  },
-  "keywords": [
-    "solana",
-    "crypto",
-    "defi",
-    "dex",
-    "amm"
-  ],
-  "author": "team@orca.so",
-  "bugs": {
-    "url": "https://github.com/orca-so/whirlpools/issues"
-  },
-  "homepage": "https://orca.so"
-}
