--- ts-sdk/actions/src/index.ts
@@ -1,8 +1,0 @@
-export * from "./config";
-export * from "./createPool";
-export * from "./decreaseLiquidity";
-export * from "./harvest";
-export * from "./increaseLiquidity";
-export * from "./pool";
-export * from "./position";
-export * from "./swap";
