--- ts-sdk/actions/src/pool.ts
@@ -1,5 +1,0 @@
-export {
-  fetchSplashPool,
-  fetchConcentratedLiquidityPool,
-  fetchWhirlpoolsByTokenPair,
-} from "@orca-so/whirlpools";
