--- ts-sdk/actions/src/position.ts
@@ -1,4 +1,0 @@
-export {
-  fetchPositionsForOwner,
-  fetchPositionsInWhirlpool,
-} from "@orca-so/whirlpools";
