+++ ts-sdk/whirlpool/src/config.ts
@@ -1,7 +1,23 @@
 import { getWhirlpoolsConfigExtensionAddress } from "@orca-so/whirlpools-client";
-import type { Address, TransactionSigner } from "@solana/kit";
-import { address, createNoopSigner, isAddress } from "@solana/kit";
+import type { Address, TransactionSigner, KeyPairSigner } from "@solana/kit";
+import {
+  address,
+  createNoopSigner,
+  isAddress,
+  createKeyPairFromBytes,
+  createSignerFromKeyPair,
+} from "@solana/kit";
 
+export {
+  setComputeUnitMarginMultiplier,
+  setJitoBlockEngineUrl,
+  setJitoTipSetting,
+  setPriorityFeeSetting,
+  setRpc,
+  setJitoFeePercentile,
+  setPriorityFeePercentile,
+  getRpcConfig,
+} from "@orca-so/tx-sender";
 /**
  * The default (null) address.
  */
@@ -169,4 +185,20 @@ export function resetConfiguration() {
   FUNDER = DEFAULT_FUNDER;
   SLIPPAGE_TOLERANCE_BPS = DEFAULT_SLIPPAGE_TOLERANCE_BPS;
   NATIVE_MINT_WRAPPING_STRATEGY = DEFAULT_NATIVE_MINT_WRAPPING_STRATEGY;
+}
+
+let _payer: KeyPairSigner | undefined;
+
+export async function setPayerFromBytes(pkBytes: Uint8Array<ArrayBuffer>) {
+  const kp = await createKeyPairFromBytes(pkBytes);
+  const signer = await createSignerFromKeyPair(kp);
+  _payer = signer;
+  return signer;
+}
+
+export function getPayer(): KeyPairSigner {
+  if (!_payer) {
+    throw new Error("Payer not set. Call setPayer() first.");
+  }
+  return _payer;
 }
